+++ src/main/java/roomescape/payment/service/dto/PaymentRequest.java
@@ -1,7 +1,9 @@
 package roomescape.payment.service.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(
-        long amount,
+        BigDecimal amount,
         String orderId,
         String paymentKey
 ) {
