+++ src/main/java/roomescape/payment/service/dto/PaymentResponse.java
@@ -1,11 +1,12 @@
 package roomescape.payment.service.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentResponse(
         String paymentKey,
         String status,
         String orderId,
-        Long totalAmount,
+        BigDecimal totalAmount,
         String method
-
 ) {
 }
