+++ src/main/java/roomescape/reservation/controller/dto/MemberReservationRequest.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MemberReservationRequest(
@@ -31,6 +33,6 @@ public record MemberReservationRequest(
 
         @NotNull(message = "결제 금액은 필수 값입니다.")
         @Positive
-        Long amount
+        BigDecimal amount
 ) {
 }
