+++ src/main/java/roomescape/reservation/controller/dto/ReservationPaymentRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ReservationPaymentRequest(
         @NotNull(message = "멤버 예약 id는 필수 값입니다.")
         @Positive
@@ -16,7 +18,6 @@ public record ReservationPaymentRequest(
 
         @NotNull(message = "결제 금액은 필수 값입니다.")
         @Positive
-        Long amount
-
+        BigDecimal amount
 ) {
 }
