+++ src/main/java/roomescape/reservation/service/dto/ThemeCreate.java
@@ -2,7 +2,9 @@ package roomescape.reservation.service.dto;
 
 import roomescape.reservation.controller.dto.ThemeRequest;
 
-public record ThemeCreate(String name, String description, String thumbnail, long price) {
+import java.math.BigDecimal;
+
+public record ThemeCreate(String name, String description, String thumbnail, BigDecimal price) {
     public static ThemeCreate from(ThemeRequest themeRequest) {
         return new ThemeCreate(themeRequest.name(), themeRequest.description(), themeRequest.thumbnail(),
                 themeRequest.price());
