+++ src/test/java/roomescape/fixture/ThemeFixture.java
@@ -2,12 +2,14 @@ package roomescape.fixture;
 
 import roomescape.reservation.domain.Theme;
 
+import java.math.BigDecimal;
+
 public class ThemeFixture {
     public static Theme getTheme1() {
-        return new Theme(null, "theme1", "description", "thumbnail.png", 15000L);
+        return new Theme(null, "theme1", "description", "thumbnail.png", BigDecimal.valueOf(15000L));
     }
 
     public static Theme getTheme2() {
-        return new Theme(null, "theme2", "description2", "thumbnail.png", 20000L);
+        return new Theme(null, "theme2", "description2", "thumbnail.png", BigDecimal.valueOf(20000L));
     }
 }
